pyproject.toml
[project]
name = "fastapiproject-film-catalog"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "art>=6.5",
    "fastapi[standard]>=0.121.1",
    "jinja2>=3.1.6",
    "pydantic-settings[yaml]>=2.12.0",
    "redis[hiredis]>=7.0.1",
    "typer>=0.20.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "coverage>=7.11.2",
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pre-commit>=4.4.0",
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "python-multipart>=0.0.20",
    "ruff>=0.14.4",
]

[tool.uv]
required-version = "~=0.8.10"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra --strict-markers"
log_cli_level = "INFO"
log_cli = false
console_output_style = "count"
markers = ['apitest: test any http api call',
    'templatetest: test any template rendering', ]
xfail_strict = true

[tool.mypy]
strict = true
plugins = ['pydantic.mypy']
ignore_missing_imports = true

[tool.ruff]
required-version = '>=0.12.1'
src = ['film-catalog']
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".venv",
]
line-length = 88
indent-width = 4
target-version = "py313"
lint.fixable = ["ALL"]
lint.select = [
    "FAST", # FastAPI
    "ANN", # flake8-annotations
    "I", # isort
    "F", # Pyflakes
    "COM", # flake8-commas
    "Q", # flake8-quotes
    "S", # flake8-bandit
    #    "ASYNC", # flake8-async
    "BLE", # flake8-blind-except
    "FBT", # flake8-boolean-trap
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T10", # flake8-debugger
    "EM", # flake8-errmsg
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "LOG", # flake8-logging
    "G", # flake8-logging-format
    "INP", # flake8-no-pep420
    "PIE", # flake8-pie
    "T20", # flake8-print
    "PYI", # flake8-pyi
    #    "PT", # flake8-pytest-style
    "RSE", # flake8-raise
    "RET", # flake8-return
    "SLF", # flake8-self
    "SIM", # flake8-simplify
    "TC", # flake8-type-checking
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "FLY", # flynt
    "C90", # mccabe
    "N", # pep8-naming
    "PERF", # Perflint
    "E", # Error
    "W", # Warning
    "PGH", # pygrep-hooks
    "PLC", # Convention
    "PLE", # Pylint Error
    "PLR", # Pylint Refactor
    "PLW", # Pylint Warning
    "UP", # pyupgrade
    "FURB", # refurb
    "RUF", # Ruff-specific rules
    "TRY", # tryceratops

]

lint.ignore = [
    "S311", # suspicious-non-cryptographic-random-usage
    "TC006", # runtime-cast-value
    "S101", # S101	assert	Use of assert detected
    "RUF001", # ambiguous-unicode-character-string (RUF001)
    "RUF002", # ambiguous-unicode-character-docstring (RUF002)
    "RUF003", # ambiguous-unicode-character-comment (RUF003)
]

[tool.ruff.lint.extend-per-file-ignores]
"stuff.py" = [
    "T20", # flake8-print (T20)
]
